+++ src/main/java/com/shop/advance/academy/yordan/petrov/git/shop/domain/models/TransactionServiceViewModel.java
@@ -17,7 +17,7 @@ public class TransactionServiceViewModel {
     private Instant dateCompleted;
     private Instant dateUpdated;
     private TransactionStatus transactionStatus = TransactionStatus.NEW;
-    private Long fee;
+    private BigDecimal fee;
     private BigDecimal amount;
     private CurrencyServiceViewModel currency;
 
@@ -96,11 +96,11 @@ public class TransactionServiceViewModel {
         this.transactionStatus = transactionStatus;
     }
 
-    public Long getFee() {
+    public BigDecimal getFee() {
         return this.fee;
     }
 
-    public void setFee(Long fee) {
+    public void setFee(BigDecimal fee) {
         this.fee = fee;
     }
